* No longer assumes the user is pi, instead auto-detecting it so the user can have any username. * Switched to udisks2 instead of pmount as pmount is very old and would only mount ntfs drives for me. * Gave the user has an option if they want to install ps3netsrv, xlinkkai, and a wifi access point, in case they don't need them to save resources and speed up the install.

+++ samba-init.sh
@@ -8,9 +8,12 @@ then
     exit
 fi
 
-#restart ps3netsrv++
-pkill ps3netsrv++
-/usr/local/bin/ps3netsrv++ -d /share
+#if /usr/local/bin/ps3netsrv++ exists
+if [ -f /usr/local/bin/ps3netsrv++ ]; then
+  #restart ps3netsrv++
+  pkill ps3netsrv++
+  /usr/local/bin/ps3netsrv++ -d /share
+fi
 
 sudo cat <<'EOF' | sudo tee /etc/samba/smb.conf
 [global]
@@ -20,7 +23,7 @@ usershare allow guests = yes
 map to guest = bad user
 allow insecure wide links = yes
 [share]
-Comment = Pi shared folder
+Comment = shared folder
 Path = /share
 Browseable = yes
 Writeable = Yes
@@ -29,11 +32,11 @@ create mask = 0777
 directory mask = 0777
 Public = yes
 Guest ok = yes
-force user = pi
+force user = userplaceholder
 follow symlinks = yes
 wide links = yes
 EOF
 
 #if you wish to create a samba user with password you can use the following:
-#sudo smbpasswd -a pi
+#sudo smbpasswd -a userplaceholder
 sudo /etc/init.d/smbd restart
